requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
hashids==1.3.1
Markdown==3.10.2
pillow==12.1.1
PyJWT==2.11.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
